Aggregate numeric features in buckets when slicing
We store histogram buckets for each numeric feature, so that we can
associate each slice with a bucket instead of a value.
The getSliceKey_ function now checks if the feature values should be
aggregated and returns the key of the appropriate bucket in the positive
case. It uses the value itself otherwise (just as before).