Add variable to configure slicing aggregation
`maxUniqueValueSlices` defines a threshold above which values should be
aggregated into buckets. Numeric features spread over a continuum
usually take many different values in the dataset (many unique values),
so that aggregation is important. On the other hand, cases in which they
assume only a few different values (or even the trivial binary [numeric]
feature) may not need aggregation.
The current threshold is set to 10 because this is the number of buckets
that the histogram provides: if we have less than 10 different values,
there's not much motivation to aggregate into 10 ranges.